switch to zurb foundation framework
This required more tweaking and fixes than is obvious
from this commit.  The grid is different and is now
properly responsive, but there was much editing of
the helper scripts that deal with image galleries
until I got it working.

Going through the above took most of the weekend,
mainly as I wanted to be sure I still needed all
the parts of my bespoke style.css file.